Add bonus right tick mark
The current way tick marks on the x axis are generated guarantees that
there won't be a tick mark higher than the max value.
In practice, however, this seems to often leave an awkward space at the
right end of the axis.

This commit adds an extra tick mark;
because of the padding, I think it should always fit.