Introduce new interval calculation algorithm
For this discussion, `interval` refers to the interval between ticks
on the x-axis.

The old tick calculation algorithm tried to fit as many ticks as
possible in the x-axis. This often resulted in somewhat awkward tick intervals,
e.g., you could have an interval of 49, when 50 would've worked just fine.

The new algorithm tries to be more flexible and choose more "round" numbers
(multiples of 5 or 10).

This issue was first flagged in #4, and PR #5 proposed one way of dealing with it.
(8b2a5dc introduced a manual work-around.)
The new algorithm is inspired by the one in #5 but is slightly more robust.