Add data quality checks when computing percentages
and deprecate `likert_percentages`

Commit 45e992a added a new `compute_percentages` argument that allows
aggregation of response counts into percentages.
The current commit adds a warning, like the one we already had for the
`plot_likert` function, if the counts vary between questions
(making percentages misleading).

With this change, the new argument can be used instead of the
`likert_percentages` function, so that function is now deprecated.
While it's supposed to be internal, I'm keeping it around in case anyone
was using it directly.