Ensure null check happens first during validation
When validating that the counted values are in the scale, the code skips
over null values, but it was performing the null check only after
looking for the value in the scale.
However, for certain null values such as pandas.NA,
the check `value in scale` produces a TypeError:
`boolean value of NA is ambiguous`.

To avoid triggering this error, this commit moves the null check to be
performed first.